refactor: split `generate-production-loaders` and use node:fs writeFile for file output
